BREAKING CHANGE: mark FFI functions as unsafe because it is.
this now requires the programmer to actually use `unsafe` before calling
any of this crate's functions. this is needed because using `unsafe {}`
inside functions not marked as unsafe is apperantly deprecated.

(or if its not, then at least you HAVE to know you're working with
unsafe code, and therefore FFI anyway. not all losses).